flaskapi-instrumentation/requirements.txt
Flask
uvicorn
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
locust